+++ CHANGE_LOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 
 - Fixed an issue where opening a chat right after agent ends the previous chat doesn't work
 - Fixed an issue where opening/refreshing a persistent chat shows the Reconnect Pane
+- Fix SSO magic code due to incompatibility of broadcast channel
 
 ## [1.2.1] - 2023-7-24